+++ batchConvert.m
@@ -1,9 +1,10 @@
-function batchConvert(inputDir, outputDir)
+function batchConvert(inputDir, outputDir, varargin)
 % batchConvert Converts original style OpenEP datasets into the new format.
 %
 % In the newer format, userdata.surface.triRep is a regular struct rather
 % than a TriRep. This is to allow for interopability with OpenEP-Py, which
-% cannot load TriRep objects.
+% cannot load TriRep objects. This function also allows options to
+% pre-compute normals.
 %
 % Usage:
 %   batchConvert(inputDir, outputDir)
@@ -15,11 +16,32 @@ function batchConvert(inputDir, outputDir)
 %               will be the same as those in inputDir.
 %
 % Author: Paul Smith (2021) (Copyright)
-% Modifications -
-%       Steven Williams (2022) Updated documentation and added notes
-%
 % SPDX-License-Identifier: Apache-2.0
 %
+% Modifications -
+%       Steven Williams (2022) Updated documentation and added notes
+%       Steven Williams (2023) Added option to pre-compute normals
+%
+% Info on Code Testing:
+% ---------------------------------------------------------------
+% batchConvert(inputDir, outputDir, 'computenormals', true)
+% ---------------------------------------------------------------
+%
+% ---------------------------------------------------------------
+% code
+% ---------------------------------------------------------------
+
+% parse input arguments
+nStandardArgs = 2;
+computeNormals = false;
+if nargin > nStandardArgs
+    for i = nStandardArgs:2:nargin-1-nStandardArgs
+        switch lower(varargin{i})
+            case 'computenormals'
+                computeNormals = varargin{i+1};
+        end
+    end
+end
 
 % Create the output folder if it does not exist
 if ~exist(outputDir, 'dir')
@@ -44,6 +66,14 @@ for i = 1:numel(allFiles)
    % Store comment about what we have done
    userdata.notes{end+1} = [date ': data set converted using batchConvert.m'];
 
+   if computeNormals
+       disp('precomputing normals')
+       [~, userdata] = getNormals(userdata);
+
+       % Store comment about what we have done
+       userdata.notes{end+1} = [date ': normals added using batchConvert.m'];
+   end
+   
    % We save as -v7 because it's faster to load in OpenEP-py than -v7.3,
    % and the saved file is significantly smaller compared to -v6 files.
    outputFile = [outputDir filesep() allFiles{i}];